LayoutTests/http/tests/webAPIStatistics/navigator-functions-accessed-data-collection-expected.txt

Tests for navigator functions accessed data collection in ResourceLoadStatistics plist by querying for all the navigator properties and dumping the entire resource load statistics map.
Resource load statistics:

Registrable domain: 127.0.0.1
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    IsScheduledForAllButCookieDataRemoval: No
    isPrevalentResource: No
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
